yhy DB

+++ conf/lsq20130411.sql
@@ -20,8 +20,8 @@ SET FOREIGN_KEY_CHECKS=0;
 -- ----------------------------
 DROP TABLE IF EXISTS `message`;
 CREATE TABLE `message` (
-  `msgId` int(11) NOT NULL,
-  `msgTypeId` int(11) default NULL,
+  `msgId` bigint(11) NOT NULL,
+  `msgTypeId` bigint(11) default NULL,
   `msgConctent` text,
   `publisherName` varchar(20) default NULL,
   `publisherPhone` varchar(20) default NULL,
@@ -42,7 +42,7 @@ CREATE TABLE `message` (
 -- ----------------------------
 DROP TABLE IF EXISTS `messagetype`;
 CREATE TABLE `messagetype` (
-  `msgTypeId` int(11) NOT NULL,
+  `msgTypeId` bigint(11) NOT NULL,
   `msgTypeName` varchar(20) default NULL,
   `msgTypeRemark` text,
   `msgTypeStatus` int(11) default '0' COMMENT '0-可用 1-不可用',
@@ -58,7 +58,7 @@ CREATE TABLE `messagetype` (
 -- ----------------------------
 DROP TABLE IF EXISTS `power`;
 CREATE TABLE `power` (
-  `powerId` int(11) NOT NULL,
+  `powerId` bigint(11) NOT NULL,
   `powerName` varchar(20) default NULL,
   `powerAction` varchar(40) NOT NULL,
   `powerStatus` int(11) default NULL,
@@ -75,7 +75,7 @@ CREATE TABLE `power` (
 -- ----------------------------
 DROP TABLE IF EXISTS `role`;
 CREATE TABLE `role` (
-  `roleId` int(11) NOT NULL,
+  `roleId` bigint(11) NOT NULL,
   `roleName` varchar(20) default NULL,
   `roleStatus` int(11) default NULL,
   `roleRemark` text,
@@ -91,8 +91,8 @@ CREATE TABLE `role` (
 -- ----------------------------
 DROP TABLE IF EXISTS `role_power`;
 CREATE TABLE `role_power` (
-  `powerId` int(11) NOT NULL,
-  `roleId` int(11) NOT NULL,
+  `powerId` bigint(11) NOT NULL,
+  `roleId` bigint(11) NOT NULL,
   PRIMARY KEY  (`powerId`,`roleId`),
   KEY `FK_role_power2` (`roleId`),
   CONSTRAINT `FK_role_power` FOREIGN KEY (`powerId`) REFERENCES `power` (`powerId`),
@@ -108,8 +108,8 @@ CREATE TABLE `role_power` (
 -- ----------------------------
 DROP TABLE IF EXISTS `user`;
 CREATE TABLE `user` (
-  `userId` int(11) NOT NULL,
-  `roleId` int(11) default NULL,
+  `userId` bigint(11) NOT NULL,
+  `roleId` bigint(11) default NULL,
   `userName` varchar(20) default NULL,
   `userPassword` varchar(20) default NULL,
   `userRemark` text,
